requirements.txt
asyncio
python-telegram-bot
openai
python-dotenv
wandb